experiments/exp1/baselines/Load_From_NFS/requirements.txt
torch
torchvision
numpy
pillow
pandas